Coded Turtle cursor movements
The Turtle is a cursor on the screen that uses a Turtle for its icon.
To draw a line, you push the Turtle forward. The screen draws a line of
the same color as the Turtle (which can be changed at any time) along
its path of movement. To draw more interesting shapes, you simply
change the direction of the Turtle. With a few more commands you can
even draw solid shapes.